+++ css/reset.css
@@ -30,6 +30,9 @@ a{
   color: inherit;
   display: block;
 }
+a:hover{
+  color: inherit;
+}
 button{
   background-color: transparent;
   border: none;
@@ -65,6 +68,7 @@ header .menu-wrap{
   align-items: center;
   width: fit-content;
   height: 100%;
+  margin: 0;
 }
 header .menu-wrap .menu{
   width: fit-content;
